+++ nest-app/src/app.module.ts
@@ -3,9 +3,9 @@ import { Module } from '@nestjs/common';
 import { TypeOrmModule } from '@nestjs/typeorm';
 import { AppController } from './app.controller';
 import { AppService } from './app.service';
-import { UsersModule } from './users/users.module';
-import { BoardsModule } from './boards/boards.module';
-import { TasksModule } from './tasks/tasks.module';
+import { UsersModule } from './resources/users/users.module';
+import { BoardsModule } from './resources/boards/boards.module';
+import { TasksModule } from './resources/tasks/tasks.module';
 import { AuthModule } from './auth/auth.module';
 
 import { CONFIG_DB } from './common/ormconfig';
