+++ nest-app/src/resources/tasks/entities/task.entity.ts
