+++ nest-app/src/resources/users/entities/user.entity.ts
@@ -1,5 +1,5 @@
-import { Task } from 'src/tasks/entities/task.entity';
 import { Column, Entity, OneToMany, PrimaryGeneratedColumn } from 'typeorm';
+import { Task } from '../../tasks/entities/task.entity';
 
 @Entity()
 export class User {
