+++ nest-app/src/resources/users/users.module.ts
@@ -1,9 +1,9 @@
 import { forwardRef, Module } from '@nestjs/common';
+import { TypeOrmModule } from '@nestjs/typeorm';
 import { UsersService } from './users.service';
 import { UsersController } from './users.controller';
-import { TypeOrmModule } from '@nestjs/typeorm';
 import { User } from './entities/user.entity';
-import { AuthModule } from 'src/auth/auth.module';
+import { AuthModule } from '../../auth/auth.module';
 
 @Module({
   imports: [forwardRef(() => AuthModule), TypeOrmModule.forFeature([User])],
