+++ nest-app/src/resources/users/users.service.ts
@@ -5,7 +5,7 @@ import { DeleteResult, Repository } from 'typeorm';
 import { CreateUserDto } from './dto/create-user.dto';
 import { UpdateUserDto } from './dto/update-user.dto';
 import { User } from './entities/user.entity';
-import { CONFIG } from '../common/config';
+import { CONFIG } from '../../common/config';
 
 @Injectable()
 export class UsersService {
