+++ nest-app/src/utils/userToResponse.ts
@@ -1,4 +1,4 @@
-import { User } from '../users/entities/user.entity';
+import { User } from '../resources/users/entities/user.entity';
 
 export const userToResponse = (
   user: User,
